UsersTable RowAction done

+++ server/handler/admin_delete_user.go
@@ -31,25 +31,27 @@ func AdminDeleteUser(c *gin.Context) {
 		return
 	}
 
-	for _, file := range user.Files() {
-		mainFilePath := filepath.Join("./uploads/", user.ID, file.ID)
-		infoFilePath := mainFilePath + ".info"
+	if len(user.Files()) > 0 {
+		for _, file := range user.Files() {
+			mainFilePath := filepath.Join("./uploads/", user.ID, file.ID)
+			infoFilePath := mainFilePath + ".info"
 
-		err := os.Remove(mainFilePath)
-		if err != nil {
-			println(err.Error())
-			c.JSON(http.StatusBadRequest, gin.H{"message": "Failed to delete file"})
-			return
+			err := os.Remove(mainFilePath)
+			if err != nil {
+				println(err.Error())
+				c.JSON(http.StatusBadRequest, gin.H{"message": "Failed to delete file"})
+				return
+			}
+
+			// Delete the info file
+			err = os.Remove(infoFilePath)
+			if err != nil {
+				println(err.Error())
+				c.JSON(http.StatusBadRequest, gin.H{"message": "Failed to delete info file"})
+				return
+			}
+
 		}
-
-		// Delete the info file
-		err = os.Remove(infoFilePath)
-		if err != nil {
-			println(err.Error())
-			c.JSON(http.StatusBadRequest, gin.H{"message": "Failed to delete info file"})
-			return
-		}
-
 	}
 
 	user, err = prisma.Client().User.FindUnique(
@@ -57,13 +59,8 @@ func AdminDeleteUser(c *gin.Context) {
 	).Delete().Exec(prisma.Context())
 
 	if err != nil {
-		// if errors.Is(err, db.ErrNotFound) {
-		// 	// success deleted user
-		// }
-	} else {
 		println(err.Error())
 		c.JSON(http.StatusBadRequest, gin.H{"message": "Failed to delete user"})
-		return
 	}
 
 	c.JSON(http.StatusOK, gin.H{"message": "success"})
